Correct errors

+++ src/main/java/globaloutbreak/controller/ControllerImpl.java
@@ -306,8 +306,8 @@ public final class ControllerImpl implements Controller {
         final Map<TypeOfInfo, String> info = new HashMap<>();
         final Optional<Region> r = this.model.getSelectedRegion();
         if (r.isPresent()) {
-            info.put(TypeOfInfo.INFETTI, Integer.toString(r.get().getNumInfected()));
-            info.put(TypeOfInfo.MORTI, Integer.toString(r.get().getNumDeath()));
+            info.put(TypeOfInfo.INFETTI, Long.toString(r.get().getNumInfected()));
+            info.put(TypeOfInfo.MORTI, Long.toString(r.get().getNumDeath()));
             info.put(TypeOfInfo.REGION, r.get().getName());
         } else {
             final InfoData infoData = this.model.getInfo();
